+++ domain/src/main/java/com/petid/domain/hospital/model/Hospital.java
@@ -2,9 +2,9 @@ package com.petid.domain.hospital.model;
 
 public record Hospital(
         Long id,
-        int sidoId,
-        int sigunguId,
-        int eupmundongId,
+        long sidoId,
+        long sigunguId,
+        long eupmundongId,
         String imageUrl,
         String address,
         String name,
